+++ pom.xml
@@ -60,13 +60,6 @@
 			</exclusions>
 		</dependency>
 		<!-- https://mvnrepository.com/artifact/io.cucumber/cucumber-java -->
-		<dependency>
-			<groupId>io.cucumber</groupId>
-			<artifactId>cucumber-java</artifactId>
-			<version>5.7.0</version>
-		</dependency>
-		
-		
 		<!-- appium dependencies -->
 		<dependency>
     		<groupId>io.appium</groupId>
