+++ qaf-blank-project-maven.iml
@@ -61,16 +61,6 @@
     <orderEntry type="library" name="Maven: velocity:velocity-dep:1.4" level="project" />
     <orderEntry type="library" name="Maven: org.json:json:20180813" level="project" />
     <orderEntry type="library" name="Maven: com.qmetry:qaf-support:3.0.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-java:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-core:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-gherkin:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-gherkin-vintage:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:tag-expressions:2.0.4" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-expressions:8.3.1" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:datatable:3.3.1" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:cucumber-plugin:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: io.cucumber:docstring:5.7.0" level="project" />
-    <orderEntry type="library" name="Maven: org.apiguardian:apiguardian-api:1.1.0" level="project" />
     <orderEntry type="library" name="Maven: io.appium:java-client:7.5.1" level="project" />
     <orderEntry type="library" scope="RUNTIME" name="Maven: org.apache.httpcomponents:httpclient:4.5.13" level="project" />
     <orderEntry type="library" scope="RUNTIME" name="Maven: org.apache.httpcomponents:httpcore:4.4.13" level="project" />
