+++ resources/application.properties
@@ -35,3 +35,6 @@ driver.name=chromeDriver
 #driver path
 webdriver.chrome.driver=driver/chromedriver
 
+# this is going to force each feature files to run independently
+selenium.singletone=Method
+
